% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsProcessing.R
\name{ConsProcessing}
\alias{ConsProcessing}
\title{reprocessing}
\usage{
ConsProcessing(GcDuoObject, DuoResults, lib, win_width = 10)
}
\arguments{
\item{DuoResults}{object obtained by procesData function}

\item{lib}{library used for identification}

\item{win_width}{window width for peaks}

\item{data}{original dataset}
}
\value{
A List
}
\description{
Process using parafac with the identified compounds obtained
}
