% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsProcessing.R
\name{ConstrainedParafac}
\alias{ConstrainedParafac}
\title{Constrained Parafac Individually}
\usage{
ConstrainedParafac(DuoPeak, GcDuoObject, DuoResults, lib_matrix)
}
\description{
Process using parafac with the identified compounds obtained for each peak
}
