% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duoIDs.R
\name{duoIDs}
\alias{duoIDs}
\title{duoIDs}
\usage{
duoIDs(DuoResults, lib, match_cutoff = 0.7, RI = NULL, RIrange = 30)
}
\arguments{
\item{DuoResults}{object obtained from processData}

\item{lib}{dataset of spectras to compare.}

\item{match_cutoff}{value to consider a correct match (range 0 to 1)}

\item{RI}{retention index of reference compounds (alkanes or FAMEs) to use, time must be in seconds}

\item{RIrange}{tolerance of RI calculation}
}
\value{
data table with the matched peaks
}
\description{
Compares spectras with library
}
