% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duoID.R
\name{getRI}
\alias{getRI}
\title{Get Retention Index}
\usage{
getRI(DuoResults, RI)
}
\arguments{
\item{DuoResults}{object result of processing}

\item{RI}{list of alkanes RT-RI}
}
\value{
DuoResult object with RI calculated
}
\description{
Calculates Kovats Retention Index based in Alkanes
}
