% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duoID.R
\name{id_rev}
\alias{id_rev}
\title{ID inverse}
\usage{
id_rev(cor_cut, spec_sample, match_cutoff, lib_matrix)
}
\arguments{
\item{cor_cut}{dot product results}

\item{spec_sample}{sample spectra}

\item{match_cutoff}{value to consider a correct match (range 0 to 1)}

\item{lib_matrix}{matrix from the library}
}
\value{
data table with the reverse similarity
}
\description{
Performs reverse similarity
}
