% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duoID.R
\name{libID}
\alias{libID}
\title{libID}
\usage{
libID(Peak2ID, lib_matrix)
}
\arguments{
\item{Peak2ID}{spectra and metadata to annotate}

\item{lib_matrix}{spectras from library to perform dot product}
}
\value{
data table with the annotated peak
}
\description{
Annotation of spectras
}
