% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_import.R
\name{library_import}
\alias{library_import}
\title{library_import}
\usage{
library_import(msp_file)
}
\arguments{
\item{lib_list}{path for the msp file/s}
}
\value{
a S3 object
}
\description{
Read libraries in msp format and transform to GcDuo
}
