% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{PlotPeak}
\alias{PlotPeak}
\title{PlotPeak}
\usage{
PlotPeak(
  GcDuoObject,
  finalGCDuo,
  peakid = "68-68-1",
  type = "eic",
  win_width = 20
)
}
\arguments{
\item{GcDuoObject}{asdasijdhjkajh}

\item{finalGCDuo}{final table with the peaks detected}

\item{peakid}{index of the selected peak to visualize}

\item{type}{select if the visualization is for tic or for eic}
}
\value{
A ggplot multiline
}
\description{
Get the graphically representation of a detected peak. Only possible to visualize one peak per time
}
