% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{procesData}
\alias{procesData}
\title{procesData}
\usage{
procesData(
  GcDuoObject,
  signalNoiseRatio1 = 10,
  signalNoiseRatio2 = 500,
  exclusionIon = NULL,
  topAbundance = 0.05,
  win_size = 3
)
}
\arguments{
\item{GcDuoObject}{raw GCxGC-MS data uploaded with `readFolderCDF`}

\item{signalNoiseRatio1}{minimum signal to noise ratio 3D}

\item{signalNoiseRatio2}{minimum signal to noise ratio 2D}

\item{exclusionIon}{mass to charge fragments to exclude in the analysis}

\item{topAbundance}{percetage of mass to change fragments to consider in the model creation}

\item{win_size}{number of modulations to include in each window}

\item{data}{defining structure of the data processing workflow}
}
\value{
A List
}
\description{
Process using parafax a section of the data defined in the dataSlot
}
