% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{procesDataSlot}
\alias{procesDataSlot}
\title{procesDataSlot}
\usage{
procesDataSlot(processingStructure)
}
\arguments{
\item{dataSlot}{defining structure of the data processing workflow}
}
\value{
A List
}
\description{
Process using parafax a section of the data defined in the dataSlot
}
