% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCDF.R
\name{readFolderCDF}
\alias{readFolderCDF}
\title{readFolderCDF}
\usage{
readFolderCDF(folderPath, modulationTime, mzRange)
}
\arguments{
\item{folderPath}{path to the folder containing the CDF to process.}

\item{modulationTime}{modulation time used in the data.}

\item{mzRange}{range values used in the m/z data acquisition.}
}
\value{
A List of arrays containing all the data in the files following the structure: |mz, time2D, time1D|
}
\description{
Read all CDF files in a folder and transform the data into a list of 3D arrays.
}
