% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duoID.R
\name{rt2ri}
\alias{rt2ri}
\title{RI calculation}
\usage{
rt2ri(rtTime, lib)
}
\arguments{
\item{rtTime}{retention time first dimension}

\item{lib}{library with RIs}
}
\value{
vector with RTs converted to RIs
}
\description{
Calculates Kovats Retention Index based in Alkanes
}
