% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_import.R
\name{spectramatrix}
\alias{spectramatrix}
\title{spectramatrix}
\usage{
spectramatrix(lis)
}
\arguments{
\item{libs}{list obtained with `getMSP` function}
}
\value{
a matrix
}
\description{
Convert the spectra list to matrix
}
