% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{SpectraView}
\alias{SpectraView}
\title{SpectraView}
\usage{
SpectraView(
  GcDuoObject,
  finalGCDuo,
  peakid = "7-1",
  lib_comp = lib,
  mzRange = c(30, 600)
)
}
\arguments{
\item{GcDuoObject}{raw GcDuo object}

\item{finalGCDuo}{processed GcDuo object}

\item{peakid}{peak ID to plot}

\item{lib_comp}{value to compare spectra with the library}

\item{mzRange}{range of m/z values to visualize}
}
\value{
An plotly graph
}
\description{
Get the graphically representation of the spectra
}
