% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeChrom2D}
\alias{visualizeChrom2D}
\title{visualizeChrom2D}
\usage{
visualizeChrom2D(GcDuoObject, samples = 1:12)
}
\arguments{
\item{GcDuoObject}{asdasijdhjkajh}

\item{sampleNum}{the index of the file/s to visualize}
}
\value{
An plotly graph
}
\description{
Get the graphically representation of the chromatogram in 2D
}
