% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualizeChrom3D}
\alias{visualizeChrom3D}
\title{visualizeChrom3D}
\usage{
visualizeChrom3D(GcDuoObject, sampleNum = 1)
}
\arguments{
\item{GcDuoObject}{asdasijdhjkajh}

\item{sampleNum}{the index of the file to visualize}

\item{data}{defining structure of the data processing workflow}
}
\value{
A ggplot contour
}
\description{
Get the graphically representation of the chromatogram in 2D. Only possible to visualize one file per time
}
