% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCDF.R
\name{readCFD}
\alias{readCFD}
\title{readCFD}
\usage{
readCFD(filePath, modulationTime)
}
\arguments{
\item{modulationTime}{modulation time used in the data}

\item{pathFolderCDF}{path to the folder containing the CDF to process.}
}
\value{
A 3 dimensional array containing all the data in the files following the structure: |sample, mz, time2D, time1D|
}
\description{
Read a single CDF file and transform the data into a 3D array.
}
