% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{SpectraView}
\alias{SpectraView}
\title{SpectraView}
\usage{
SpectraView(
  GcDuoObject,
  finalGCDuo,
  peakid = "7-1",
  lib_comp = T,
  mzRange = c(30, 600)
)
}
\arguments{
\item{GcDuoObject}{asdasijdhjkajh}

\item{sampleNum}{the index of the file/s to visualize}
}
\value{
An plotly graph
}
\description{
Get the graphically representation of the spectra
}
